+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Added side hitboxes to blocks
+- Changed renderer to only render blocks within view of the player
 
 # 0.0.20_2
 *2024-08-15 01:40*